+++ front-end/src/App.js
@@ -1,8 +1,8 @@
 import "./App.scss";
-import Body from "./components/Body/Body";
-import Footer from "./components/Footer/Footer";
-import Header from "./components/Header/Header";
-import Sidebar from "./components/Sidebar/Sidebar";
+import Body from "./components/Body";
+import Footer from "./components/Footer";
+import Header from "./components/Header";
+import Sidebar from "./components/Sidebar";
 
 function App() {
   return (
