--- front-end/src/components/Body/Current/Current.jsx
@@ -1,15 +1,0 @@
-import React from "react";
-import "./style.scss";
-
-const Current = (props) => {
-  return (
-    <>
-      <div className="Current">
-        <p>
-          <span>Home</span>/<span>Users</span>
-        </p>
-      </div>
-    </>
-  );
-};
-export default Current;
