+++ front-end/src/components/Body/Current/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import "./style.scss";
+
+export default function Current() {
+  return (
+    <div className="Current">
+      <p>
+        <span>Home</span>/<span>Users</span>
+      </p>
+    </div>
+  );
+}
