+++ front-end/src/components/Footer/index.jsx
